backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.9
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.23
flask-login==0.6.2
oauth2client==4.1.3
google-auth==2.22.0
requests==2.31.0
python-dotenv==1.0.0